requirements.txt
transformers
einops
requests
filetype
pillow
torch
torchvision
streamlit
